Move src/pages/components to src/components
