+++ src/components/input/search-bar.tsx
@@ -12,8 +12,8 @@ import {
 } from "@chakra-ui/react";
 import { useState, SetStateAction } from "react";
 import { useTranslation } from "next-i18next";
-import Trie from "../../../filter/trie";
-import AutocompleteFilter from "../../../filter/autocomplete";
+import Trie from "../../filter/trie";
+import AutocompleteFilter from "../../filter/autocomplete";
 
 const SearchBar = ({
   baseWidth,
