+++ src/components/nav/nav-items.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import Link from "next/link";
 import { useTranslation } from "react-i18next";
-import styles from "../../../../styles/Home.module.css";
+import styles from "../../../styles/Home.module.css";
 import { Button } from "@chakra-ui/react";
 
 export default function NavItems() {
