+++ src/pages/search/index.tsx
@@ -10,8 +10,8 @@ import { tag } from "../../types/arweave";
 import { glossetaSearchResult } from "../../types/glosseta-lookup-item";
 import { Result } from "./result";
 import { UnavailableResult } from "./unavailable-result";
-import PageLayout from "../components/layout/page";
-import SearchBar from "../components/input/search-bar";
+import PageLayout from "../../components/layout/page";
+import SearchBar from "../../components/input/search-bar";
 import ApiError from "./api-error";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { termFilter } from "../../filter/termConfig";
